requirements.txt
# Core NLP and ML
# Using CPU-only torch to reduce image size and build time
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
accelerate>=0.25.0
openai>=1.3.0
google-generativeai>=0.3.0
groq>=0.4.2

# Search and Database
elasticsearch==8.11.0
opensearch-py>=2.4.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
pydantic>=2.5.0
python-multipart
pydantic-settings>=2.1.0

# Task Queue
celery>=5.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
biopython>=1.81
nltk>=3.8.0
spacy>=3.7.0

# Web Scraping and Data Fetching
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tenacity>=8.2.0
loguru>=0.7.0
tqdm>=4.66.0

# Monitoring
prometheus-client>=0.19.0

# Testing (Optional for Prod)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
python-telegram-bot>=20.0
flask>=3.0.0
